# 8315041: Optimize Java to C string conversion by avoiding double copy