--- a/src/java.base/share/classes/java/lang/foreign/SegmentAllocator.java
+++ b/src/java.base/share/classes/java/lang/foreign/SegmentAllocator.java
@@ -125,11 +125,20 @@ public interface SegmentAllocator {
         Objects.requireNonNull(charset);
         Objects.requireNonNull(str);
         int termCharSize = StringSupport.CharsetKind.of(charset).terminatorCharSize();
-        byte[] bytes = str.getBytes(charset);
-        MemorySegment segment = allocateNoInit(bytes.length + termCharSize);
-        MemorySegment.copy(bytes, 0, segment, ValueLayout.JAVA_BYTE, 0, bytes.length);
+        MemorySegment segment;
+        int length;
+        if (StringSupport.bytesCompatible(str, charset)) {
+            length = str.length();
+            segment = allocateNoInit(length + termCharSize);
+            StringSupport.copyToSegmentRaw(str, segment, 0);
+        } else {
+            byte[] bytes = str.getBytes(charset);
+            length = bytes.length;
+            segment = allocateNoInit(bytes.length + termCharSize);
+            MemorySegment.copy(bytes, 0, segment, ValueLayout.JAVA_BYTE, 0, bytes.length);
+        }
         for (int i = 0 ; i < termCharSize ; i++) {
-            segment.set(ValueLayout.JAVA_BYTE, bytes.length + i, (byte)0);
+            segment.set(ValueLayout.JAVA_BYTE, length + i, (byte)0);
         }
         return segment;
     }