@mcimadamore 2023-08-25
Numbers are as follows:

Before:

```
Benchmark                         (size)  Mode  Cnt   Score   Error  Units
ToCStringTest.panama_writeString       5  avgt   30  51.716 ? 1.489  ns/op
ToCStringTest.panama_writeString      20  avgt   30  51.704 ? 0.881  ns/op
ToCStringTest.panama_writeString     100  avgt   30  57.848 ? 2.402  ns/op
ToCStringTest.panama_writeString     200  avgt   30  60.734 ? 3.613  ns/op
```

After:

```
Benchmark                         (size)  Mode  Cnt   Score   Error  Units
ToCStringTest.panama_writeString       5  avgt   30  45.227 ? 0.329  ns/op
ToCStringTest.panama_writeString      20  avgt   30  47.015 ? 0.491  ns/op
ToCStringTest.panama_writeString     100  avgt   30  47.571 ? 1.447  ns/op
ToCStringTest.panama_writeString     200  avgt   30  48.630 ? 0.954  ns/op
```

Of course, the bigger the string size, the bigger the speedup. But, even for small strings, we get a nice boost.