@mcimadamore 2023-08-28
> @mcimadamore Great work! Do you plan to similarly remove the double copy necessary for C to Java String conversion?

I thought about that. That conversion consists of three steps:

1. find the length of the C string
2. bulk copy the string bytes into a new byte[]
3. create a new string from the bytes (which copies the bytes again)

In principle, it should be possible to avoid the allocation in (3) by passing the `byte[]` computed in (2) directly to a private string constructor (of course, depending on the charset being used). But, to do this efficiently, we need to be able to quickly determine as to whether all the bytes in the array are positive (for `UTF8` and `ASCII`). In the Java to C conversion we do that by using StringCoder::hasNegatives` which is backed by a vectorized JVM intrinsics. Since this method only works for `byte[]` input, not for `MemorySegment`, we have a problem there: if scanning the string bytes takes too long, we're in a pickle, because the time we save for the allocation will go in the new scanning.

Ideally, you'd like to check for negatives _while_ doing (1), but the logic for (1) is already quite complex. It is something we might look at in the future.